+++ Legion/Utils.h
@@ -4,6 +4,7 @@ namespace Utils
 	bool ShouldWriteFile(string Path);
 	string GetTimestamp();
 	string GetDate();
+	string Vector3ToHexColor(Math::Vector3 vec);
 };
