+++ Legion/src/Assets/material.cpp
@@ -1,0 +1,197 @@
+#include "pch.h"
+#include "RpakLib.h"
+#include "Path.h"
+#include "Directory.h"
+
+void RpakLib::BuildMaterialInfo(const RpakLoadAsset& Asset, ApexAsset& Info)
+{
+	auto RpakStream = this->GetFileStream(Asset);
+	IO::BinaryReader Reader = IO::BinaryReader(RpakStream.get(), true);
+
+	RpakStream->SetPosition(this->GetFileOffset(Asset, Asset.SubHeaderIndex, Asset.SubHeaderOffset));
+
+	MaterialHeader MatHeader = Reader.Read<MaterialHeader>();
+
+	RpakStream->SetPosition(this->GetFileOffset(Asset, MatHeader.NameIndex, MatHeader.NameOffset));
+
+	string MaterialName = Reader.ReadCString();
+
+	if (ExportManager::Config.GetBool("UseFullPaths"))
+		Info.Name = MaterialName;
+	else
+		Info.Name = IO::Path::GetFileNameWithoutExtension(MaterialName).ToLower();
+
+	Info.Type = ApexAssetType::Material;
+	Info.Status = ApexAssetStatus::Loaded;
+
+	uint32_t TexturesCount = 0;
+
+	switch (Asset.Version) {
+	case RpakGameVersion::Apex:
+		TexturesCount = (MatHeader.UnknownOffset - MatHeader.TexturesOffset) / 8;
+		break;
+	case RpakGameVersion::Titanfall:
+	case RpakGameVersion::R2TT: // unverified but should work
+		TexturesCount = (MatHeader.UnknownTFOffset - MatHeader.TexturesTFOffset) / 8;
+		break;
+	}
+	Info.Info = string::Format("Textures: %i", TexturesCount);
+}
+
+void RpakLib::ExportMaterial(const RpakLoadAsset& Asset, const string& Path)
+{
+	RMdlMaterial Material = this->ExtractMaterial(Asset, Path, false, false);
+	string OutPath = IO::Path::Combine(Path, Material.MaterialName);
+
+	if (!Utils::ShouldWriteFile(OutPath))
+		return;
+
+	IO::Directory::CreateDirectory(OutPath);
+
+	(void)this->ExtractMaterial(Asset, OutPath, true, true);
+}
+
+RMdlMaterial RpakLib::ExtractMaterial(const RpakLoadAsset& Asset, const string& Path, bool IncludeImages, bool IncludeImageNames)
+{
+	RMdlMaterial Result;
+
+	auto RpakStream = this->GetFileStream(Asset);
+	IO::BinaryReader Reader = IO::BinaryReader(RpakStream.get(), true);
+
+	RpakStream->SetPosition(this->GetFileOffset(Asset, Asset.SubHeaderIndex, Asset.SubHeaderOffset));
+
+	MaterialHeader MatHeader = Reader.Read<MaterialHeader>();
+
+	RpakStream->SetPosition(this->GetFileOffset(Asset, MatHeader.NameIndex, MatHeader.NameOffset));
+
+	Result.MaterialName = IO::Path::GetFileNameWithoutExtension(Reader.ReadCString());
+
+	List<ShaderResBinding> PixelShaderResBindings;
+
+	bool shadersetLoaded = Assets.ContainsKey(MatHeader.ShaderSetHash);
+
+	if (shadersetLoaded)
+	{
+		RpakLoadAsset ShaderSetAsset = Assets[MatHeader.ShaderSetHash];
+		ShaderSetHeader ShaderSetHeader = ExtractShaderSet(ShaderSetAsset);
+
+		uint64_t PixelShaderGuid = ShaderSetHeader.PixelShaderHash;
+
+		if (ShaderSetAsset.AssetVersion <= 11)
+		{
+			PixelShaderGuid = ShaderSetHeader.OldPixelShaderHash;
+		}
+
+		if (Assets.ContainsKey(PixelShaderGuid))
+		{
+			PixelShaderResBindings = ExtractShaderResourceBindings(Assets[PixelShaderGuid], D3D_SHADER_INPUT_TYPE::D3D_SIT_TEXTURE);
+		}
+		else {
+			g_Logger.Warning("Shaderset for material '%s' referenced a pixel shader that is not currently loaded. Unable to associate texture types.\n", Result.MaterialName.ToCString());
+		}
+	}
+
+	g_Logger.Info("\nMaterial Info for '%s'\n", Result.MaterialName.ToCString());
+	g_Logger.Info("> ShaderSet: %llx (%s)\n", MatHeader.ShaderSetHash, shadersetLoaded ? "LOADED" : "NOT LOADED");
+
+	const uint64_t TextureTable = (Asset.Version == RpakGameVersion::Apex) ? this->GetFileOffset(Asset, MatHeader.TexturesIndex, MatHeader.TexturesOffset) : this->GetFileOffset(Asset, MatHeader.TexturesTFIndex, MatHeader.TexturesTFOffset);
+	uint32_t TexturesCount = (Asset.Version == RpakGameVersion::Apex) ? 0x10 : 0x11;
+
+	// we're actually gonna ignore the hardcoded value for apex materials
+	if (Asset.Version == RpakGameVersion::Apex)
+	{
+		TexturesCount = (MatHeader.UnknownOffset - MatHeader.TexturesOffset) / 8;
+		g_Logger.Info("> %i textures:\n", TexturesCount);
+	}
+
+	// These textures have named slots
+	for (uint32_t i = 0; i < TexturesCount; i++)
+	{
+		RpakStream->SetPosition(TextureTable + ((uint64_t)i * 8));
+
+		uint64_t TextureHash = Reader.Read<uint64_t>();
+		string TextureName = "";
+		bool bOverridden = false;
+		bool bNormalRecalculate = false;
+
+		if (TextureHash != 0)
+		{
+
+			TextureName = string::Format("0x%llx%s", TextureHash, (const char*)ImageExtension);
+
+			if (PixelShaderResBindings.Count() > 0 && i < PixelShaderResBindings.Count())
+			{
+				string ResName = PixelShaderResBindings[i].Name;
+				if (!ExportManager::Config.GetBool("UseTxtrGuids"))
+				{
+					TextureName = string::Format("%s_%s%s", Result.MaterialName.ToCString(), ResName.ToCString(), (const char*)ImageExtension);
+				}
+				bOverridden = true;
+
+				if (ResName == "normalTexture")
+					bNormalRecalculate = true;
+			}
+
+			if (Asset.Version == RpakGameVersion::Apex)
+				g_Logger.Info(">> %i: 0x%llx - %s\n", i, TextureHash, bOverridden ? TextureName.ToCString() : "(no assigned name)");
+
+			switch (i)
+			{
+			case 0:
+				Result.AlbedoHash = TextureHash;
+				Result.AlbedoMapName = TextureName;
+				break;
+			case 1:
+				Result.NormalHash = TextureHash;
+				Result.NormalMapName = TextureName;
+				break;
+			case 2:
+				Result.GlossHash = TextureHash;
+				Result.GlossMapName = TextureName;
+				break;
+			case 3:
+				Result.SpecularHash = TextureHash;
+				Result.SpecularMapName = TextureName;
+				break;
+			case 4:
+				Result.EmissiveHash = TextureHash;
+				Result.EmissiveMapName = TextureName;
+				break;
+			case 5:
+				Result.AmbientOcclusionHash = TextureHash;
+				Result.AmbientOcclusionMapName = TextureName;
+				break;
+			case 6:
+				Result.CavityHash = TextureHash;
+				Result.CavityMapName = TextureName;
+				break;
+			}
+		}
+
+		// Extract to disk if need be
+		if (IncludeImages && Assets.ContainsKey(TextureHash))
+		{
+			RpakLoadAsset& Asset = Assets[TextureHash];
+			// Make sure the data we got to is a proper texture
+			if (Asset.AssetType == (uint32_t)AssetType_t::Texture)
+			{
+				ExportTexture(Asset, Path, IncludeImageNames, bOverridden ? TextureName : string(), bNormalRecalculate);
+			}
+		}
+	}
+
+	g_Logger.Info("\n");
+
+	return Result;
+}
+
+std::unique_ptr<Assets::Texture> RpakLib::BuildPreviewMaterial(uint64_t Hash)
+{
+	if (!this->Assets.ContainsKey(Hash))
+		return nullptr;
+
+	auto& Asset = this->Assets[Hash];
+	RMdlMaterial Material = this->ExtractMaterial(Asset, "", false, false);
+
+	return this->BuildPreviewTexture(Material.AlbedoHash);
+}
